Fix testing for TRandomPoisson (#1058)

+++ root/dataframe/test_stringfiltercolumn.C
@@ -20,7 +20,7 @@ void getTracks(FourVectors& tracks) {
    auto nPart = R.Poisson(5);
    tracks.clear();
    tracks.reserve(nPart);
-   for (int i = 0; i < nPart; ++i) {
+   for (int i = 0; i < static_cast<int>(nPart); ++i) {
       double px = R.Gaus(0,10);
       double py = R.Gaus(0,10);
       double pt = sqrt(px*px +py*py);
